+++ packages/manager/apps/dedicated/client/app/components/dedicated-cloud/dashboard/tiles/options/components/nsxt/nsxt.component.js
@@ -1,0 +1,11 @@
+import template from './nsxt.html';
+import controller from './nsxt.controller';
+
+export default {
+  bindings: {
+    currentService: '<',
+  },
+  name: 'ovhManagerPccDashboardOptionsNsxt',
+  template,
+  controller,
+};