+++ packages/manager/core/application/src/hooks/index.ts
@@ -1,0 +1,4 @@
+export * from './useAuthentication';
+export * from './useEnvironment';
+export * from './useLogger';
+export * from './useNavigation';