+++ packages/manager/core/application/src/hooks/useNavigation.tsx
@@ -1,0 +1,8 @@
+import { useShell } from '..';
+
+export function useNavigation() {
+  const shell = useShell();
+  return shell.navigation;
+}
+
+export default useNavigation;