+++ packages/manager/modules/iplb/src/frontends/iplb-frontends.controller.js
@@ -31,10 +31,20 @@ export default class IpLoadBalancerFrontendsCtrl {
     this.loading = true;
     this.IpLoadBalancerFrontendService.getFrontends(
       this.$stateParams.serviceName,
-    ).then((results) => {
-      this.loading = false;
-      this.frontends = results;
-    });
+    )
+      .then((frontends) => {
+        this.frontends = frontends.map((frontend) => {
+          const { protocol, ssl } = frontend;
+          const displayProtocol =
+            ssl && protocol === 'http' ? 'https' : protocol;
+          set(frontend, 'displayProtocol', displayProtocol);
+
+          return frontend;
+        });
+      })
+      .finally(() => {
+        this.loading = false;
+      });
   }
 
   loadFarm(frontend) {