+++ packages/manager/modules/pci/src/projects/project/data-processing/jobs/job-details/job-dashboard/job-dashboard.component.js
@@ -8,9 +8,7 @@ export default {
   bindings: {
     job: '<',
     projectId: '<',
-    metricsToken: '<',
     terminateJob: '<',
-    showMetrics: '<',
     showBillingConsole: '<',
     browseObjectStorage: '<',
     goBack: '<',