+++ packages/manager/modules/pci/src/projects/project/data-processing/jobs/job-details/job-dashboard/job-dashboard.controller.js
@@ -1,4 +1,4 @@
-import { find, unzip } from 'lodash';
+import { find } from 'lodash';
 import {
   formatDuration,
   getDataProcessingUiUrl,
@@ -6,20 +6,17 @@ import {
 import {
   DATA_PROCESSING_STATUS_TO_CLASS,
   DATA_PROCESSING_STATUSES,
-  METRICS_REFRESH_INTERVAL,
+  JOB_REFRESH_INTERVAL,
   JOB_TYPE_JAVA,
   JOB_TYPE_PYTHON,
   DATA_PROCESSING_TRACKING_PREFIX_FULL,
 } from '../../../data-processing.constants';
-import { WARP10_URL } from './job-dashboard.constants';
 
 export default class {
   /* @ngInject */
   constructor(
     $scope,
-    $resource,
     $timeout,
-    $uibModal,
     $window,
     CucCloudMessage,
     dataProcessingService,
@@ -41,35 +38,7 @@ export default class {
     this.atInternet = atInternet;
     this.containerId = null;
     this.pollTimer = null;
-    this.metricsTimer = null;
     this.formatDuration = formatDuration;
-    // setup metrics retrieval
-    this.warp10 = $resource(
-      WARP10_URL,
-      {},
-      {
-        query: {
-          method: 'POST',
-          transformRequest: [],
-          isArray: true,
-        },
-      },
-    );
-    // setup metrics container
-    this.metrics = {
-      totalMemory: {
-        data: [],
-        labels: [],
-      },
-      activeTasks: {
-        data: [],
-        labels: [],
-      },
-      blockManagerDiskUsed: {
-        data: [],
-        label: [],
-      },
-    };
   }
 
   $onInit() {
@@ -83,7 +52,7 @@ export default class {
         this.containerId = container.id;
       }
     });
-    // start metrics retrieval
+    // start job retrieval
     this.pollData();
   }
 
@@ -94,18 +63,15 @@ export default class {
   }
 
   /**
-   * Query metrics from OVH metrics backend
-   * If job is still running, we query each METRICS_REFRESH_INTERVAL to update charts
+   * Query job from OVH backend
+   * If job is still running, we query each JOB_REFRESH_INTERVAL to update charts
    */
   pollData() {
-    this.queryMetricsTotalMemory();
-    this.queryMetricsActiveTasks();
-    this.queryMetricsDiskUsed();
     this.queryJob();
     if (!this.job.endDate) {
       this.pollTimer = this.$timeout(
         () => this.pollData(),
-        METRICS_REFRESH_INTERVAL,
+        JOB_REFRESH_INTERVAL,
       );
     }
   }
@@ -138,93 +104,6 @@ export default class {
       .then((job) => {
         this.job = job;
       });
-  }
-
-  /**
-   * Retrieve JVM total memory usage for current job from metrics backend
-   */
-  queryMetricsTotalMemory() {
-    const d = this.computeDates();
-    this.warp10
-      .query(
-        `[ '${
-          this.metricsToken.data.token
-        }' 'spark_jvm_memory_usage' { 'qty' 'used' 'mem_type' 'total' 'job-id' '${
-          this.job.id
-        }' } '${d.startDate.toISOString()}' '${d.endDate.toISOString()}' ] FETCH SORT [ SWAP [ 'executor-id' ] reducer.sum ] REDUCE`,
-      )
-      .$promise.then((series) => {
-        if (series.length > 0 && series[0][0] && 'v' in series[0][0]) {
-          let data = series[0][0].v;
-          data = unzip(data);
-          this.metrics.totalMemory = {
-            labels: data[0].map((o) => moment(o / 1e3).format('hh:mm:ss')),
-            data: data[1].map((o) => parseInt(o / 1e6, 10)),
-          };
-        }
-      });
-  }
-
-  /**
-   * Retrieve BlockManager disk usage for current job from metrics backend
-   */
-  queryMetricsDiskUsed() {
-    const d = this.computeDates();
-    this.warp10
-      .query(
-        `[ '${
-          this.metricsToken.data.token
-        }' 'spark_block_manager' { 'qty' 'diskSpaceUsed_MB' 'type' 'disk' 'job-id' '${
-          this.job.id
-        }' } '${d.startDate.toISOString()}' '${d.endDate.toISOString()}' ] FETCH SORT`,
-      )
-      .$promise.then((series) => {
-        if (series.length > 0 && series[0][0] && 'v' in series[0][0]) {
-          let data = series[0][0].v;
-          data = unzip(data);
-          this.metrics.blockManagerDiskUsed = {
-            labels: data[0].map((o) => moment(o / 1e3).format('hh:mm:ss')),
-            data: data[1].map((o) => Math.floor(o / 1e6)),
-          };
-        }
-      });
-  }
-
-  /**
-   * Retrieve Spark activate tasks for current job from metrics backend
-   */
-  queryMetricsActiveTasks() {
-    const d = this.computeDates();
-    this.warp10
-      .query(
-        `[ '${
-          this.metricsToken.data.token
-        }' 'spark_executor' { 'qty' 'activeTasks' 'type' 'threadpool' 'job-id' '${
-          this.job.id
-        }' } '${d.startDate.toISOString()}' '${d.endDate.toISOString()}' ] FETCH SORT`,
-      )
-      .$promise.then((series) => {
-        if (series.length > 0 && series[0][0] && 'v' in series[0][0]) {
-          let data = series[0][0].v;
-          data = unzip(data);
-          this.metrics.activeTasks = {
-            labels: data[0].map((o) => moment(o / 1e3).format('hh:mm:ss')),
-            data: data[1].map((o) => Math.floor(o)),
-          };
-        }
-      });
-  }
-
-  /**
-   * Checks if at least one chart has data
-   */
-  hasCharts() {
-    return (
-      this.metrics.totalMemory.data.length +
-        this.metrics.activeTasks.data.length +
-        this.metrics.blockManagerDiskUsed.data.length >
-      0
-    );
   }
 
   /**
@@ -261,14 +140,6 @@ export default class {
       getDataProcessingUiUrl(this.job.region, this.job.id),
       '_blank',
     );
-  }
-
-  onMetricsClick() {
-    this.atInternet.trackClick({
-      name: `${DATA_PROCESSING_TRACKING_PREFIX_FULL}::jobs::job-details::dashboard::monitor`,
-      type: 'action',
-    });
-    this.showMetrics();
   }
 
   onObjectContainerClick() {