+++ packages/manager/modules/pci/src/projects/project/kubernetes/details/service/service.component.js
@@ -25,6 +25,7 @@ const component = {
     addOidcProvider: '<',
     updateOidcProvider: '<',
     removeOidcProvider: '<',
+    isKubernetesInstalling: '<',
   },
   template,
   controller,