+++ packages/manager/tools/sao-ovh-manager-app/template/src/index.js
@@ -1,20 +1,26 @@
 import 'script-loader!jquery'; // eslint-disable-line
 import 'core-js/stable';
 import 'whatwg-fetch';
-import { bootstrapApplication } from '@ovh-ux/manager-core';
+import 'regenerator-runtime/runtime';
+
+import { isTopLevelApplication } from '@ovh-ux/manager-config';
 import { defineApplicationVersion } from '@ovh-ux/request-tagger';
+
+import { useShellClient } from '@ovh-ux/shell';
 
 defineApplicationVersion(__VERSION__);
 
-bootstrapApplication('<%= name %>').then((environment) => {
-  if (environment.getMessage()) {
-    displayMessage(environment.getMessage(), environment.getUserLanguage());
+useShellClient('<%= name %>').then((shellClient) => {
+  if (!isTopLevelApplication()) {
+    shellClient.ux.startProgress();
   }
-
-  import(`./config-${environment.getRegion()}`)
-    .catch(() => {})
-    .then(() => import('./app.module'))
-    .then(({ default: startApplication }) => {
-      startApplication(document.body, environment);
-    });
+  shellClient.environment.getEnvironment().then((environment) => {
+    environment.setVersion(__VERSION__);
+    import(`./config-${environment.getRegion()}`)
+      .catch(() => {})
+      .then(() => import('./app.module'))
+      .then(({ default: startApplication }) => {
+        startApplication(document.body, shellClient);
+      });
+  });
 });