+++ packages/components/ng-shell-tracking/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.2](https://github.com/ovh/manager/compare/@ovh-ux/ng-shell-tracking@0.2.1...@ovh-ux/ng-shell-tracking@0.2.2) (2023-06-12)
+
+**Note:** Version bump only for package @ovh-ux/ng-shell-tracking
+
+
+
+
+
+## [0.2.1](https://github.com/ovh/manager/compare/@ovh-ux/ng-shell-tracking@0.2.0...@ovh-ux/ng-shell-tracking@0.2.1) (2023-05-31)
+
+**Note:** Version bump only for package @ovh-ux/ng-shell-tracking
+
+
+
+
+
 # [0.2.0](https://github.com/ovh/manager/compare/@ovh-ux/ng-shell-tracking@0.1.6...@ovh-ux/ng-shell-tracking@0.2.0) (2023-03-06)