+++ packages/manager/apps/carrier-sip/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.15](https://github.com/ovh/manager/compare/@ovh-ux/manager-carrier-sip-app@2.7.14...@ovh-ux/manager-carrier-sip-app@2.7.15) (2023-06-12)
+
+**Note:** Version bump only for package @ovh-ux/manager-carrier-sip-app
+
+
+
+
+
+## [2.7.14](https://github.com/ovh/manager/compare/@ovh-ux/manager-carrier-sip-app@2.7.13...@ovh-ux/manager-carrier-sip-app@2.7.14) (2023-05-31)
+
+**Note:** Version bump only for package @ovh-ux/manager-carrier-sip-app
+
+
+
+
+
 ## [2.7.13](https://github.com/ovh/manager/compare/@ovh-ux/manager-carrier-sip-app@2.7.12...@ovh-ux/manager-carrier-sip-app@2.7.13) (2023-05-04)
 
 **Note:** Version bump only for package @ovh-ux/manager-carrier-sip-app