+++ packages/manager/apps/office/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.18](https://github.com/ovh/manager/compare/@ovh-ux/manager-office-app@0.7.17...@ovh-ux/manager-office-app@0.7.18) (2023-06-12)
+
+**Note:** Version bump only for package @ovh-ux/manager-office-app
+
+
+
+
+
+## [0.7.17](https://github.com/ovh/manager/compare/@ovh-ux/manager-office-app@0.7.16...@ovh-ux/manager-office-app@0.7.17) (2023-05-31)
+
+**Note:** Version bump only for package @ovh-ux/manager-office-app
+
+
+
+
+
 ## [0.7.16](https://github.com/ovh/manager/compare/@ovh-ux/manager-office-app@0.7.15...@ovh-ux/manager-office-app@0.7.16) (2023-04-20)
 
 **Note:** Version bump only for package @ovh-ux/manager-office-app