+++ packages/manager/apps/pci/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.7.48](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@3.7.47...@ovh-ux/manager-pci-app@3.7.48) (2023-06-12)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
+## [3.7.47](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@3.7.46...@ovh-ux/manager-pci-app@3.7.47) (2023-06-12)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
+## [3.7.46](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@3.7.45...@ovh-ux/manager-pci-app@3.7.46) (2023-06-07)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
+## [3.7.45](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@3.7.44...@ovh-ux/manager-pci-app@3.7.45) (2023-05-31)
+
+**Note:** Version bump only for package @ovh-ux/manager-pci-app
+
+
+
+
+
 ## [3.7.44](https://github.com/ovh/manager/compare/@ovh-ux/manager-pci-app@3.7.43...@ovh-ux/manager-pci-app@3.7.44) (2023-05-30)
 
 **Note:** Version bump only for package @ovh-ux/manager-pci-app